split up data generation, data inspection, and DataLoader/Dataset code
