+++ main.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import numpy as np
 import yaml
 
-from signal_analysis.datasets import (
+from signal_analysis.generate_data import (
     generate_pretraining_data,
     generate_training_data_from_rp,
 )
